Define wells

Define wells in the aquifer with required parameters.

define_wells(wells_df = NULL, ...)

Arguments

wells_df
A data.frame (or tibble, or sf with point geometry) containing properties of wells. If wells_df is an sf object, the coordinates for each well are extracted and the sf geometry is removed.

...
Named vector inputs to be added to the wells object. Names that replicate columns in wells_df supplant the columns in wells_df, except wID and well_type which are generated automatically. The length of inputs should generally be the same as the number of wells. However, if an input is repeated for all wells, it can be input with length 1.

Value

This function returns a data.frame containing the columns:

wID: Well ID

Q: Pumping rate

R: Radius of influence

diam: Diameter of the well

x, y: Cartesian coordinates

well_type: Pumping (Q < 0), Injection (Q > 0), or Non-operational (Q = 0) well

These columns are followed by any other columns desired for the wells.

Examples

define_wells(data.frame(wID=c(3,NA,9)),Q=rep(0,3))

#> # A tibble: 3 x 8
#>     wID     Q     R  diam     x     y well_type       well_image
#>   <int> <dbl> <dbl> <dbl> <dbl> <dbl> <fct>           <chr>     
#> 1     3     0    NA    NA    NA    NA Non-operational Actual    
#> 2    10     0    NA    NA    NA    NA Non-operational Actual    
#> 3     9     0    NA    NA    NA    NA Non-operational Actual    

wells_df <- data.frame(x=1:3,y=1:3,Q=11:13,diam=0.5,id=1001:1003)
wells <- define_wells(wells_df)
print(wells) # prints a tibble of the wells

#> # A tibble: 3 x 9
#>     wID     Q     R  diam     x     y well_type well_image    id
#>   <int> <dbl> <dbl> <dbl> <dbl> <dbl> <fct>     <chr>      <int>
#> 1     1    11    NA   0.5     1     1 Injection Actual      1001
#> 2     2    12    NA   0.5     2     2 Injection Actual      1002
#> 3     3    13    NA   0.5     3     3 Injection Actual      1003

define_wells(wells_df,Note="note",Name=c("N1","N2","N3"))

#> # A tibble: 3 x 11
#>     wID     Q     R  diam     x     y well_type well_image    id Note  Name 
#>   <int> <dbl> <dbl> <dbl> <dbl> <dbl> <fct>     <chr>      <int> <chr> <chr>
#> 1     1    11    NA   0.5     1     1 Injection Actual      1001 note  N1   
#> 2     2    12    NA   0.5     2     2 Injection Actual      1002 note  N2   
#> 3     3    13    NA   0.5     3     3 Injection Actual      1003 note  N3   

define_wells(wells_df,R=1e4,Q=-21:-23) # include R and replace Q in wells_df

#> # A tibble: 3 x 9
#>     wID     Q     R  diam     x     y well_type well_image    id
#>   <int> <int> <dbl> <dbl> <dbl> <dbl> <fct>     <chr>      <int>
#> 1     1   -21 10000   0.5     1     1 Pumping   Actual      1001
#> 2     2   -22 10000   0.5     2     2 Pumping   Actual      1002
#> 3     3   -23 10000   0.5     3     3 Pumping   Actual      1003